Update ClassGraph.kt
a bug in the get function of the ClassGraph class. The bug occurs when the keepPackages and ignorePackages patterns match the same class name. In this case, the class is incorrectly added to the entryPoints set, even if it should be ignored.
solution:
To fix the bug, we need to modify the condition in the get function to account for cases where both keepPackages and ignorePackages patterns match the class name. We can do this by adding an additional check to ensure that the class name is not matched by the ignorePackages pattern.

Signed-off-by: Astitv Bajpai <141858691+astitvabajpai@users.noreply.github.com>