Honour createdResources cache retention for local build cache
The value configured for `createdResources.removeUnusedEntriesAfterDays()`
will now be used if `DirectoryBuildCache.removeUnusedEntriesAfterDays()` is not set.

In order to detect if DirectoryBuildCache.setRemoveUnusedEntriesAfterDays() has been called,
the default value has been changed to -1. This is a minor breaking change and we may way a
better way to determine if the user has configured this value.

The overall default of 7 days remains, since this is the default retention period for
created resources.