Move transforms output directory into Gradle-versioned cache
Previously, the transform outputs were cached in a was a cross-version
shared directory (caches/transforms-4). However, the entries in this cache
are bound to a specific Gradle version, and would never be shared between
versions.

With this change, the transform outputs are held in a gradle-versioned
cache directory (caches/<gradle-version>/transforms). This makes logical
sense and will simplify the management and cleanup of these caches.