Remove old unused 'caches/transforms-X' directories
Previously, old 'caches/transforms-X' directories would live forever:
the latest version would be purged of unused entries, but the prior versions
would never be removed even if no current Gradle version was using them.

Now when performing cleanup for `caches/modules-2`, we now also check for
unused `caches/transforms-X` directories. This cannot be tied to the cleanup
of the transforms cache, since that now lives in `caches/<gradle-version>/transforms`
and is inherently tied to the Gradle version.

Fixes #28178