Move getConfiguration and getConfigurationName to sub-interfaces
We want to (initially) remove the concept of configurations from local components. Eventually we want to
remove the concept of configurations from all non-ivy components, but that is a future goal.

We move the definitions of these methods from the top-level ComponentGraphResolveMetadata to the sub interfaces of
LocalComponentMetadata and ModuleComponentGraphResolveMetadata. Upcoming work will remove these methods entirely
from the local component interfaces.

This change allows us to better reason about where these methods are used for local components.