Remove usages of ComponentResolveMetadata
The primary motivation here is to simplify LocalComponentMetadata (now renamed to LocalComponentGraphResolveMetadata),
by reducing the type hierarchy. This will make it much easier to heavily modify the way local component metadata works.

ComponentResolveMetadata was a legacy type, previously used to represent a component when resolving artifacts. Now,
the modern type, ComponentArtifactResolveMetadata, has been propogated to all prior places that used the legacy type.

All methods on the removed type have been moved to its remaining subinterface, ExternalComponentResolveMetadata. This
subinterface is not heavily used, but removing it would cause more work than was desired for this change.