Support += for ConfigurableFileCollections in Kotlin DSL
This is a backwards-compatible change. While FileCollection supported
plus for a while, += can only fall back to that when the target is a var
of a FileCollection type, which doesn't trigger new += behavior. A var
of ConfigurableFileCollection type could not be +=, because the result
of + is a FileCollection, which isn't assignable back to the variable.