Move lazy logic from local component metadata to local component state
Previously, we had two levels of caching of variants between component metadata and component state.
The component metadata cached variant metadata instances and the component state cached component state
instances. In practice, this double caching layer is not necessary, as the cached methods on the metadata
were only called by the cache state methods.

Making this change clears the way to allow us to properly lock on project state when calculating these
values lazily. With the double caching we previously had, this was very difficult to do correctly and
in a performant manner.

Furthermore, this makes the local component metadata immutable and stateless, thus updating it to
follow the contract set forth by ComponentGraphResolveMetadata.

This commit does not itself add the proper locking mechanisms, but makes it possible to do so with few changes.