Do not always calculate available variants when resolving task dependencies
This data is only used for reporting tasks and is almost never needed. Only capture the data if it is used.
(and in reality, this data isn't even needed in the in-memory resolution result builder either, since it is
captured and sourced from the full resolution result anyways)

Also move tests that verify tests cannot share a test type into the test report aggregation integ test.
We no longer realize all variants in the current project's component when we resolve a configuration in
that component. This means we do not call the code that detects duplicate test types when we run a test.

This is a known problem, and the solution is to build a more robust artifact reselection mechanism that can
support multiple tests with the same test type. For now, we move these tests into a test that does trigger the
error.