Add central configuration for build-cache retention
Adds a new `CacheConfigurations.buildCache` setting which can be used to
configure the retention period for local build cache entries. This parallels
the other configuration settings for `downloadedResoures`, `distribution-jars`, etc.

The existing `DirectoryBuildCache.removeUnusedEntriesAfterDays` is deprecated.

If the default value for `DirectoryBuildCache.removeUnusedEntriesAfterDays` is not
changed by user code, then we use the value for `CacheConfigurations.buildCache` instead.
This results in no behaviour change, unless the user is explicitly setting the new
`CacheConfigurations.buildCache`.