Resolve a symlink Groovy JAR location in classpath (#28664)
In certain environments the classpath can contain a symlinked location
for the Groovy libraries that can be named different than the target
location. This causes the logic matching the name of the JAR file to
fail although the target location is preserving the actual expected
name.

Resolve the symlink if any is found to eliminate the failure in this
code path.

Signed-off-by: Alex Volanis <avolanis@amazon.com>