application-plugin: Make the wrapper scripts REUSE compliant
The Free Software Foundation's REUSE tool [1] ensures that all files in
a repository have license and copyright information declared. Gradle
projects that have the wrapper committed and want to be REUSE compliant
so far needed to explicitly declare the license for the scripts as part
of the `.reuse/dep5` file. Avoid that need by adding SPDX identifier
tags to the wrapper scripts, see [2].

[1]: https://reuse.software/
[2]: https://reuse.software/spec/#copyright-and-licensing-information

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>