Switch from `plugins {...}` block to software types in Declarative DSL
Remove support for `plugins { ... }` in declarative files. Along with that, the support
is removed for the top-level dependencies and, temporarily, for typesafe project accessors.

Migrate the Declarative DSL integration tests to use software types.

Refactor evaluation schema construction in a way that project schema is available without
touching project instances and can be produced once all the software types have been registered.