Execute Gradle in parent context in Windows scripts
Windows script `gradlew.bat` executes Gradle within its own, batch context.
This causes every locally declared variable to be visible to underlying `java`
process, *potentially* leaking information about used options and classpath.

To mitigate this, we jump to a non-existent label, forcing `cmd.exe` to stop
processing the script, yet run the remaining commands within the current line
in parent (most likely command-line when invoking `gradlew` directly) context.
This introduces a harmless side-effect in form of corrupting the current title
of the window `cmd.exe` was invoked in, thus it is reset to the default value.

This provides a few side benefits:

  - no more 'Terminate batch job (Y/N)?' prompt
  - `call ` and `call` are used to directly set exit code, resolving #4784
  - both sh and `cmd.exe` scripts become synchronized in terms of mechanism
    used to execute JVM as well as environment sanitization

Issue: #10463

Signed-off-by: mataha <mataha@users.noreply.github.com>