Do not execute AutoRun commands when invoking `cmd.exe`
When invoking `cmd.exe`, if `/d` was not specified on the command line,
the shell will execute AutoRun commands (specified by certain registry
variables) regardless of whether it's an interactive shell or not (e.g.
like zsh sources `.zshenv` for all shell instances). This can result in
a significant slowdown if AutoRun consists of lengthy preparatory code.

Signed-off-by: mataha <mataha@users.noreply.github.com>