Split testing-base into separate infrastructure project
Previously :testing-base included both code that ran in the gradle daemon and in the test worker.
This meant that test worker code could accidentally depend on code that it should not -- like
things from :core. Previously, developers had to be sure to not accidentally depend on these
restricted dependencies, and the build did not enforce that boundry.

This commit splits :testing-base into a separate :testing-base-infrastructure, where the infrastructure
project contains all code that runs on workers and does not depend on projects that are not intended
to run in workers.