Don't run ToolchainsParallelActionExecutionCrossVersionSpec with embedded executer
`ToolchainsParallelActionExecutionCrossVersionSpec` has been flaky for a few months.
Looking at the history, it only happens on embedded executer with
TAPI `current -> current` execution.

It's possible that some shared state is broken by other tests
in `embeddedCrossVersionTest`: the test itself is good, but
the way we run it is wrong.

This PR disables it for embedded executer.