Specify explict `--java-version` in `MultiProjectJvmApplicationInitIntegrationTest`
The test cases in `MultiProjectJvmApplicationInitIntegrationTest` attempt
to initialize a project using Java 21 as the default value.
This leads to the integration tests provisioning a local JDK,
which we typically avoid on CI due to it causing flakiness.

This PR explicitly specifies the `--java-version` to match the Java version of the test JVM,
thereby preventing the download of additional JDKs.