Prune some test worker dependencies
The eventual goal is for ForkedTestClasspathFactory to not need a hard-coded list of gradle modules
for worker classpaths. We should be able to generate that automatically based on the real dependencies
of the test worker's runtime classpath.

This commit begins to clean up the dependency graph of the test classpath, removing some dependencies
on Gradle modules and external dependencies that don't necessarily need to be there. There is still
quite some work to do, but this starts the process.