Update source code to be compatible with Java 21 compiler
This does not target the code to 21 bytecode, but updates the code to compile when using the 21 comiler and targeting 8 bytecode.

- Most notably, the 'this-escape' warning was added. Over 200 files are affected and this pattern seems to be relatively widely
used across the codebase. Furthermore, the Gradle API itself seems to promote this pattern in some instances. We should think
about how we can work to migrate away from the this-escape warning. But, for now, given that the exist code _does_ work, and
the number of files affected, all instances of the warning have been suppressed -- no matter if they are simple or difficult to resolve.
- All constructors for URL have been deprecated, and the compiler complains about them -- even though we did not change the
target bytecode version. These usages have been migrated to use the URI constructor instead. This may only affect Kotlin code.
- A 'String.indent' Kotlin extension function needed to be renamed since there is now a function on String with that name
introduced in JDK 12.
- The compiler now complains about using @see within curly braces. These usages have been mostly updated to use @see as a tag. In one case,
the linked class was unused. The class was removed, and the javadoc updated to reflect the actual code.