Extract convention processing out of interpretation sequence step implementations
Introduce resolution result handlers in the evaluator which react to the presence of
the corresponding feature flags in the step.

This simplifies the step objects and makes them easier to transfer and reuse.

Introduce an abstraction layer between the `SoftwareTypeRegistry` and the evaluator's
code handling conventions, so that the latter can replace the implementation
with a standalone one when it runs outside the Gradle process.