Clean up uses of GFileUtils.tail
This method had only one use in the DaemonDiagnostics and overall looks
quite ad-hoc, so it makes sense to move it closer to its client, which
also allows us to use Guava and Java 8 goodies for simplification.

DaemonDiagnostics runs in the client process which may have a different
default encoding, so we're using UTF-8 encoding to write the daemon log.
This code is not cross-versioned, so backward and forward compatibility
considerations do not apply.

Nullability issues were addressed along the way.