Prevent Gradle classpath from leaking onto compile classpath
When no `-classpath` parameter is passed to the Java compiler, it will attempt to infer the compile
classpath by looking at the `java.class.path` system property.

When executed without a forking compiler, this system property points to the Gradle distribution classpath.

In cases where a compilation had no dependencies, it would have an empty compile classpath. We handled
this by not passing a `-classpath` parameter. As such, classes were able to compile against Gradle
implementation classes even if they did not declare a `gradleApi` dependency.

This fixes that bug by always passing a `-classpath` flag to the compiler, even if the classpath is empty.

See Location.java in the compiler for more details on how the classpath is handled when the flag is not provided:
https://github.com/openjdk/jdk/blob/9ac8d05a2567fbf65b944660739e5f8ad1fc2020/src/jdk.compiler/share/classes/com/sun/tools/javac/file/Locations.java#L774-L776

See the original commit where this bug was likely introduced:
https://github.com/gradle/gradle/commit/23c2caee464a0088ce4e16a4a840a6c84a558e27