Rework the DOM impl types + some things around resolution
Provide implementations that can be rebuilt without
the need to carry the language tree elements around.

Instead, keep the DOM node -> language tree mapping
in a document data container.

For a resolved DOM, return the document along with
the resolution container, as otherwise it won't be possible
to properly query the resolution container.