Allow constructing `DocumentWithResolution` outside `declarative-dsl-core`
There is no real harm in that. Even with the internal constructor, one
would be able to mix up the document with a non-matching resolution
container and pass them to some of our APIs.

With this allowed, we can be a little more flexible in the other modules
around `declarative-dsl-core` where we potentially need to make these
pairs.