Properly publish public API
Use configurations and components to avoid manually writing POM XML

We use configurations to define the distribution 'source' where we plan to extract the public API from
We also have separate configurations to define the external depenencies we want to expose for the published library.

We package most of the logic in a plugin, simplifying the buildscript.
We further reduce the plugins we are applying. Previously applied the 'distribution.packaging' plugin, however this
project is not a distribution and did not need all that extra configuration.