Filter conflict candidates that have been deselected
When a conflict is detected, we do not immediately resolve it. Instead, we remove any transitive nodes of the conflicting
nodes from the graph. This is becase we do not want to continue selecting transitive dependencies of the conflicting nodes,
as one of them will be removed and the transtive work will be for nothing. We do not resolve the conflict until after the
queue is empty, as pending nodes in the queue might actually remove one of the conflicting nodes from the graph -- effectively
making the conflict a non-issue.

However, in that case, where other changes to the graph would otherwise change a conflict (remove nodes that did conflict,
so that there is only one or none remaining), we failed to detect that the conflicting node was deselected. Instead, prior code
continued with conflict resolution, potentially picking a "winner" that was already deselected by other means.

This commit ensures we only perform conflict resolution on nodes that are still present in the graph at the time of conflict resolution.
This ensures we do not erroneously pick a winner for the conflict that would otherwise not be selected

Simplify filtering logic