Ensure that build service processes its queue of events before closing
The listener registry handles closing of a build service that registered
in it by unsubscribing it from events and waiting for the queue of
unprocessed events to be drained.

Previously, it was possible for the build service to receive build
events after it was closed. That caused data races and could re-create
a destroyed service, violating the contract, as the re-created service
is never closed.

There should be no observable changes for the most of the builds. The
blocking part moved from buildFinished callback of the listener registry
to buildFinished callback of the service registry, but the time spent
waiting should be roughly the same. In most cases these callbacks happen
at about same time (when all builds of the main tree finish), except for
the nested build trees used by precompiled script plugin accessor
generator and GradleBuild task. These two now block when their nested
build tree is being disposed.

Additionally, the services registered in the nested build trees
unsubscribe early, when the tree is disposed, and not linger until the
end of the build.