Get rid of `o.g.internal.agents` package
There was only one remaining `InstrumentingClassLoader` interface left.
It cannot be moved to the instrumentation-agent-services yet because it
is used by the `base-services` classloader implementation, which require
Java 6 compatibility.

This is not a perfect state to be in, because this interface is only in
play when the agent is used (which implies running in daemon process
with Java 8), so using it in Java 6-compatible process is probably a
mistake. However, fixing this is likely very involved.