Allow Module instances to expose component ID
Clean up difference betwen owner and component ID, allowing the domain object
context to expose the owner while the module only exposes the component identity.

In order to this, we needed to move the Module implementation into dependency management
so the impls have access to the module component ID constructors.

Eventually, we should also move DependencyMetadataProvider and Module interfaces to dependency-management,
but at the moment, core still depends on these types.