Use secure algorithm for configuration cache encryption
Configuration cache encryption currently uses the Electronic Codebook (ECB) mode of AES.
This mode is generally regarded as insecure. See:
https://en.wikipedia.org/wiki/Block_cipher_mode_of_operation#Electronic_codebook_(ECB)

In general, for any algorithm to be secure, it must include a random initialization vector --
a string of bytes which must never be reused as an IV with the same key. This ensures that the
same plaintext never maps to the same ciphertext when encrypted more than once.

There are a number of AES modes that satisfy this constraint. However, CTR and GCM modes can be parallelized.
Most well-known modes, such as CBC, CFB, and OFB, rely on some form of serial chaining of blocks
during encryption and decryption.

This commit migrates CC from the AES/ECB/PKCS5PADDING transformation to the AES/GCM/NoPadding transformation.