Use CTR as default encryption algorithm
GCM for some reason appears to be much slower than CTR. This should not be the case and the
performance of these algorithms should be comparable. We should try to investigate why GCM
is so slow here, though for now we are moving forward with this effort since CTR is much
preferred over ECB.

One reason for the performance hit _might_ be the tag length used in GCM, which it currently
configured to its max value. However, it seems unlikely that this is the cause. We can try to
experiment with using the bouncycastle provider by default to see if that also affects performance,
or we should try running the benchmarks on different machines which might offer hardware acceleration
while our performance test machines on CI might not offer GCM hardware acceleration.