Apply review comments
Do not use 'either or' wording in error message, implying Gradle does not know the true answer. Instead,
suggest the user run the :outgoingVariants command to check themselves.

Update javadoc, class names
Remove extra getArtifactVariants() method that could otherwise be removed

Update upgrade guide to be in correct place in file and to be more clear with how this situation occurs.