Document reproducer for #27947
This issue was fixed by #29193, however fixing this issue was a side-effect of the change made there.

The linked changes updated DefaultLocalComponentGraphResolveState so that it reset the list of graph selection
candidates when reevaluate is called.

This reproducer can be verified by commenting out initCalculatedValues in DefaultLocalComponentGraphResolveState,
and observing that this test fails.

The reason this was caused is because the root variant is stored in a separate field compared to the consumable variants.
In the reproducer, a single declared dependency is inherited by the root variant and a consumable variant, both of which
live in the same graph.

When we resovle the graph for build dependencies, we cache values of both of these varaints (root and consumable) without
the exclude. Then beforeResolve runs, adding the exlude, calling reevaluate. The bug here was that only one of these cached
values was reset. This caused one edge to the direct dependency to not have an exclude, causing the transitive edge to
be present in the final result.

The fix was to ensure we always invalidate both caches for both the root and consumable variants.