Strengthen root node selection deprecation
The original goal of this deprecation is to prevent the root node from being selected as
a variant. This way, for a given resolution, a variant can either be intended for consumption
as a variant, or act as a root node, but not both.

Previously, we detected this scenario by checking if there were any incoming edges to the root node
by the time the graph was finished resolving.

We update this to trigger a deprecation if an incoming edge is _ever_ added to the root node in any circumstance.

This is stronger than the original deprecation, since this also handles cases where edges temporarilly target the root
node but are then later retargeted to another node for some reason.