More precisely verify tooling API deprecation warnings
Previously, we would only check for the presence of a generalized deprecation warning message, not the specific
deprecation warning emitted.

This is better than what we had, but there are still some shortcomings:
1. The expected deprecation warnings and build outputs are tightly tied to the withConnection methods on
ToolingApiSpecification. Any tests directly calling withConnection on the toolingApi fixture itself
will not properly reset captured standard output and deprecation expectations across connections.
Integ tests handle this by having the executor handle deprecations. That would be very similar to having
the toolingApi test fixture track deprecations -- though that was more work than desired here.
2. A single connection can run multiple 'long-running' tasks. The proper soution here would be to have each
ToolingApiConfigurableLauncher track its own stdout and stderr instead of merging it with each other launcher's
stdout and stderr. It becomes difficult to track which deprecations came from which gradle launch with the way
it is currently
3. Emitted deprecations are only checked if the test calls a method that verifies deprecation warnings. These
are assertHasBuildSuccessfulLogging and similarly named methods. Many tests do not call these sorts of methods
and therefore do not get deprecation checks. In the future, we should work to add tooling to verify that each
test actually performs some kind of output verification if they ran a build.