Consider producer schema during graph validation
Previously, we ignored the producer component's schema when determining if the variants in that component
were mutally compatible. This break the case where a project depends on another but the consuming project
does not have the same attribute schema as the producing project. In this case, the build could fail claiming
that the producer component has incompatible variants when in fact it is the consumer project that does
not know that those variants are actually mutually compatible.

This commit uses the producer AND consumer schema when verifying that variants are mutually compatible.