Use ResultAssertion for smoke tests
Updates smoke tests to use the same infrastructure for scanning build outputs for deprecation logs as the integ tests and tooling api tests.
We also update adhoc executors created with GradleDistribution#execter to enable deprecation warnings by default
Furthermore, we make it a warning (by default) to expect deprecation warnings when deprecation warning detection has been disabled. For smoke tests, we allow this since it is very common to disable deprecation warnings for old AGP or KGP versions.
We also add the capability for ResultAssertion to have "maybe expected" deprecations. This is leveraged by smoke tests and will be leveraged in future changes.

As a result of these changes, a number of tests needed to be updated, particularly the smoke tests.
The ResultAssertion infrastructure detects many more warnings compared to the previous deprecation detection infrastructure used by smoke tests.
forwardOutput was removed from most usages of the smoke test gradle runner, since the runner forwards output by default.
A smoke test for Kotlin precompiled script plugins was deleted since we already have one in the cross version tests, and needing to support multiple gradle versions in the smoke tests was complicated and only required by this test.