Only java conventions get java-library applied
This restricts the application of the java-library plugin in our buildlogic convention plugins to only project explicitly named as java projects.

The projectHealth check points out that a project shouldn't have both java-library and kotlin("jvm") plugins applied to it.

This adjusts the convention plugin hierarchy to avoid adding java-library to conventions that are jvm language agnostic by creating a new "jvm-library" base convention.  It also renames the "instrumented-project" convention to indicate it is only for java language projects.