Allow tooling API tests to fail before emitting BUILD FAILED
In some cases, for example when using the tooling API with an invalid JVM, an exception is thrown and
BUILD FAILED is not printed. In these cases, do not require the specific build output and accept that
an exception being thrown is enough evidence to verify that the build has indeed failed.