Consider nodes selected NOT by VAR when checking for capability conflicts
You should not be allowed to select 2 variants of the same project when selecting configurations explicitly by name, this case was not being checked properly.

This might have been an overzealous, slight performance optimization.  But configurations by name are an unusual case, so most of the time there won't be many of them, and this shouldn't have a performance impact.

In the case when there ARE many configurations selected by name, this WILL have a performance impact, but this is unavoidable for the sake of correctness.