Resolve a number of bugs resulting in corrupt dependency graphs
This commit, among other things, resolves a number of bugs that cause corrupt dependency graphs. These
corrupt graph manifest to the user in a number of ways, including cryptic error messages, missing
artifacts, or silent errors in the resolved graph. These sorts of errors are unacceptable.

The most notable issue resolved by this commit is to filter the nodes involved in a capability
conflict before resolving that conflict. Previously, a capability conflict resolver would be able
to select a non-selected node as the winner of a capability conflict. Since this node would not
otherwise be part of the graph, this caused serious issues in the final resolution.

We include a number of other fixes, including:
- Not filtering conflict candidates at the time of registering the conflict, as the graph state can
change by the time the conflict on that capability is resolved
- Ensuring all capability conflict resolution causes have reasons
- Short-circuit resolving a node's dependencies if it is involved in a conflict
- Fix a bug where the pending dependencies of a module is not properly tracked when the module is replaced by another.
Previously, these hard edges were not properly tracked and we mistakenly removed edges that we assumed were constraints.
This surfaced as many errors, including missing first-level dependencies, silent incorrect graphs, and corrupt serialized
resolution result errors.
- Remove ResolutionConflictTracker, which was based on incomplete logic and is no longer needed after these changes
- Collapse capability conflicts on the same capability
- Add documentation throughout. Adjust method names and variable names as necessary