Improve Configuration#withDependencies documentation
The prior documentation suggested to use the method for mutating dependencies. Given changes to the
way dependency resolution handles mutation of configurations, using withDependencies for this use case
is no longer recommended. Instead, for adding dependencies to a configuration lazily, users should use
addLater or addAllLater on DependencySet or DependencyConstraintSet. Further, instead of mutating
dependencies with withDependencies, users should add constraints that influence the resolution of
other dependencies (if they're added to the graph).

We describe this in detail in the javadoc for this method. We also update the user guide to remove
the recommendation of using withDependencies to mutate dependencies and instead add snippets showing
the proper way to lazily add dependencies and influence dependency version selection.

Finally, we update the build to avoid withDependencies where possible.