Deprecate running Gradle on Java versions < 17
This PR deprecates running the Gradle daemon on JVM versions less than 17.

This causes every build run on JVM < 17 to emit a deprecation warning. We handle this in integration tests
, smoke tests, and tooling API tests by blanket-ignoring all deprecation warnings with this pattern. Then,
we add a switch to _not_ ignore the deprecation warning.

For integration tests, we add tests to verify that we fail, emit a deprecation, and don't emit a deprecation
for situations where we run the build with an unsupported, deprecated, and non deprecated JVM.

For tooling API tests and GradlRunner tests, we do the same, except we test the situations where the client
is the same as the daemon version and when they are different.

For smoke test, we simply ignore the deprecation warning.

ProjectBuilde also emits a deprecation warning when the current JVM is not comaptible with Java 17 ,.

All of these new tests are supported by improvements to AvailableJavaHomes, and the preconditions that verify
certain JVMs are available on the target machines. We have checks to
- Verify an unsupported JVM is available
- Verify a depreacted JVM is avilable
- Verify a non deprecated JVM is avilable

Additionally, we can verify the _current_ jvm satisfies these conditions.

Along with the deprecation and new tests, we improve startup of the tooling API by
honoring the JAVA_HOME environment variable set by the launcher.

We also detect if the JVM that will be used by the tooling API is different than the
launcher JVM, and if so, do not start the build in debug mode -- even if requested