Do not use JavaVersion when determining java version for toolchain
If this code is loaded in the tooling API client, it will use the client's JavaVersion implementation, which may not
be able to represent older java versions. This leads us to report the incorrect java version for a discovered toolchain
when determining if it is compatible.

Instead, we use the raw version string when constructing the JavaLanguageVersion, skipping JavaVersion entirely, causing
us to report the correct java version for a toolchain