Do not try to detect different daemon JVMs when launching with embedded
Remove the logic to detect when the daemon jvm will be different than the launcher jvm. This not
necessarily a bad idea but we should think harder to see if this is the logic we want.

For now, we will leave this out of the deprecation PR to avoid any behavior changes