Create ProjectIdentity
We create ProjectIdentity, with the goal of fully encapsulating the identity of a Project in a single type.
This type is able to completely define a Project, for use in a ProjectDependency, ProjectComponentIdentifier,
and ProjectComponentSelector, without any additional services.

A ProjectDependency can now be easily converted to a ProjectComponentSelctor without accessing the backing
Project object. Many other conversions between project-identity-like types are now much easier too. Existing
code frequently accessed Default implementation fields, which were specially stored for the purpose of converting
between types. Now, ProjectIdentity makes it easy.

We made attempts to simplify construction of these types wherever possible, and standardize them on ProjectIdentity.